pyproject.toml
[project]
name = "ave-test-task"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "factory-boy>=3.3.3",
    "fastapi>=0.124.0",
    "httpx>=0.28.1",
    "phonenumbers>=9.0.20",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pytest>=9.0.2",
    "pytest-mock>=3.15.1",
    "pytest-mock-resources[redis]>=2.12.4",
    "redis>=7.1.0",
    "uvicorn>=0.38.0",
]

[tool.pytest.ini_options]
python_files = ["test_*.py", "*_test.py", "*_tests.py"]
testpaths = ["app/tests"]
